bullets can be shot from player, bullet spawn point does not follow player as it rotates

+++ src/entities/asteroid.hpp
@@ -9,7 +9,7 @@ using namespace std;
 
 class Asteroid : public Node {
 public:
-    Asteroid(Vector2 position, int width, int height, int rotation_degrees = 0);
+    Asteroid(Vector2 position, int width, int height, double rotation_degrees = 0);
 
     void input(SDL_Event *event) override;
     void physics_process(float delta) override;
